+++ include/converters.h
@@ -31,7 +31,7 @@ public:
         LabelDef *labels[128];
         getLabels(registryID, labels, num);
 
-        for (size_t i = 0; i < num; i++)
+        for (int i = 0; i < num; i++)
         {
             char *input = data;
             input += labels[i]->offset + 3;
@@ -46,7 +46,7 @@ public:
         int num = def->dataSize;
         double dblData = NAN;
         Serial.print("Converting from:");
-        for (size_t i = 0; i < num; i++)
+        for (int i = 0; i < num; i++)
         {
             Serial.printf(" 0x%02x ", data[i]);
         }
